Get post社区功能提供前端更多数据(username，community_name)等等

+++ dao/mysql/community.go
@@ -34,7 +34,7 @@ func GetCommunityList() (*[]models.Community, error) {
 }
 
 // GetCommunityDetailByID 得到某个社区详情
-func GetCommunityDetailByID(id uint64) (*models.CommunityDetail, error) {
+func GetCommunityDetailByID(id int) (*models.CommunityDetail, error) {
 	var cd models.CommunityDetail
 	sqlStr := `SELECT community_id, community_name, introduction, create_time FROM community WHERE community_id = ?`
 
@@ -46,6 +46,5 @@ func GetCommunityDetailByID(id uint64) (*models.CommunityDetail, error) {
 		}
 		return nil, err
 	}
-
 	return &cd, nil
 }
